+++ src/main/java/com/example/fashionforecastbackend/global/config/SecurityConfig.java
@@ -1,12 +1,11 @@
 package com.example.fashionforecastbackend.global.config;
 
-import com.example.fashionforecastbackend.oauth2.service.CustomOAuth2UserService;
+import com.example.fashionforecastbackend.oauth2.service.Impl.CustomOAuth2UserServiceImpl;
 import java.util.Collections;
 import java.util.List;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.security.config.Customizer;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
@@ -18,10 +17,10 @@ import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
 @EnableWebSecurity
 @Configuration
 public class SecurityConfig {
-	private final CustomOAuth2UserService customOAuth2UserService;
+	private final CustomOAuth2UserServiceImpl customOAuth2UserServiceImpl;
 
-	public SecurityConfig(CustomOAuth2UserService customOAuth2UserService) {
-		this.customOAuth2UserService = customOAuth2UserService;
+	public SecurityConfig(CustomOAuth2UserServiceImpl customOAuth2UserServiceImpl) {
+		this.customOAuth2UserServiceImpl = customOAuth2UserServiceImpl;
 	}
 
 	@Bean
@@ -33,7 +32,7 @@ public class SecurityConfig {
 			.httpBasic(AbstractHttpConfigurer::disable)
 					.oauth2Login((oauth2) -> oauth2
 							.userInfoEndpoint((userInfoEndpointConfig -> userInfoEndpointConfig
-									.userService(customOAuth2UserService))));
+									.userService(customOAuth2UserServiceImpl))));
 		http
 			.authorizeHttpRequests(auth -> auth
 				.anyRequest().permitAll());
